fix: optimize backend loading and web form
- Load DBs once in cloud function so that it is not recomputed every invocation
- Reordered main webpage form components for clarity and to stop a bug where pressing the enter key would erase the query and replace with a prefilled one.
